+++ changelog.md
@@ -3,6 +3,8 @@
 ## Upcoming
 * Fixed Better hurt by target removing the "alert others" feature
 * Hopefully fixed Dark Art Witches crashing with some other mod
+* Fixed mobs no longer running from creepers
+* Fixed launching creepers no longer launching vertical
 
 ## 2.0.3
 * Mobs that run from explosions will now also dismount to run away